% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospatsF.R
\name{ospatsF}
\alias{ospatsF}
\title{Ospats Function from Ospats-package}
\usage{
ospatsF(
  data = NULL,
  dRange = NULL,
  nCycles = NULL,
  dStart = NULL,
  ClusterStart = NULL,
  dMaxrun = NULL,
  dRSquare = NULL,
  dStrata = NULL,
  initialTemperature = NULL,
  coolingRate = NULL,
  debug = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{data matrix with columns \link{x y pred var} in that specific order}

\item{dRange}{Range parameter of the exponential correlation model}

\item{nCycles}{Number of iterations per one run of the optimisation algorithm}

\item{dStart}{choose between kMeans (0) or CumrootSquare(1) or external (3)}

\item{ClusterStart}{external for dStart == 3 (Saby Input)}

\item{dMaxrun}{Number of independent (apart from same initial states) repeated optimisation runs}

\item{dRSquare}{R^2 parameter for adjusting for "regression-to-the-mean" (see the paper)}

\item{dStrata}{Number of strata to have}

\item{initialTemperature}{simulated annealing parameter}

\item{coolingRate}{simulated annealing parameter}

\item{debug}{Useful during development for troubleshooting issues}

\item{verbose}{Print runtime diagnostics}
}
\description{
Copy-paste of the original code (with small edits)
}
\details{
These are the refactoring changes to the original code to minimize the footprint:
\enumerate{
\item The call \code{fastPdist2(Ar = as.matrix(z), Br = as.matrix(z)))^2} is now  \code{outer(z, z, "-")^2}
\item The call \code{fastVsum(Ar = as.matrix(s2))} is now \code{outer(s2, s2, "+")}
\item The call \code{fastPdist2(Ar = xy, Br = xy)} is now \code{as.matrix(dist(xy)) }
}

Other than that the code has not been touched. Only this documentation has been added.
}
\section{Original documentation}{

\subsection{opstats algorithm:}{

As described in de Gruijter et al. (2015) {Optimizing Stratification and Allocation for
Design-Based Estimation of Spatial Means Using Predictions with Error}
}

\subsection{What does it do:}{

Given a map (for instance a digital soil map) with prediction variance this function will derive a
spatial stratifcation somewhere on the spectrum between and including compact geographical stratification
(high uncertainty) and clusters based on the quantiles of the data (low uncertainty).
}
}

