% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospatsF-julia.R
\name{ospatsF_julia}
\alias{ospatsF_julia}
\title{Ospats Julia Version in R}
\usage{
ospatsF_julia(
  data,
  dRange,
  dRSquare,
  dMaxrun,
  nCycles,
  dStrata,
  verbose = FALSE
)
}
\arguments{
\item{data}{data matrix with columns [x y pred var] in that specific order}

\item{dRange}{Range parameter of the exponential correlation model}

\item{dRSquare}{R^2 parameter for adjusting for "regression-to-the-mean" (see the paper)}

\item{dMaxrun}{Number of independent (apart from same initial states) repeated optimisation runs}

\item{nCycles}{Number of iterations per one run of the optimisation algorithm}

\item{dStrata}{Number of strata to have}

\item{verbose}{Print runtime diagnostics}
}
\description{
Translate the Julia version of Ospats to R
}
\details{
The original Julia code:

https://github.com/jjdegruijter/ospats/ , file "ospatsmr"

The original code uses global variables; This version mimics the input naming of the original ospatsF.

This code only computes the stratification, and not the Neyman allocation and sampling in "Section 6" of the Julia code,
nor does the code include "Section 7." which writes the results into a file.
}
