% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospatsF-julia-refactored.R
\name{ospatsF_julia_ref}
\alias{ospatsF_julia_ref}
\title{Ospats Julia Version in R}
\usage{
ospatsF_julia_ref(
  x,
  nstrata = 3,
  niter_outer = 3,
  niter = 100,
  verbose = FALSE,
  rsquared = 1,
  covmodel_range,
  Cov
)
}
\arguments{
\item{x}{data, with variables \link{x y pred var}}

\item{nstrata}{number of starta to consider}

\item{niter_outer}{Number of independent runs of the optimisation algorithm}

\item{niter}{Number of iterations per one run of the optimisation algorithm}

\item{verbose}{Print runtime diagnostics?}

\item{rsquared}{The R^2 in the paper (default: 1)}

\item{covmodel_range}{Range parameter for assumed exponential correlation model}

\item{Cov}{Optional, overrides the covariance matrix calculation using exp-correlation. No checks with data variances.}
}
\description{
Refactor the Julia version of Ospats to R
}
\details{
The original Julia code:

https://github.com/jjdegruijter/ospats/ , file "ospatsmr"

The original code uses global variables; This version mimics the input naming of the original ospatsF.

This code only computes the stratification, and not the Neyman allocation and sampling in "Section 6" of the Julia code,
nor does the code include "Section 7." which writes the results into a file.

This "_ref" is a refactored version, similar to ospatsF_ref, notably the input names change.
}
\seealso{
\code{\link[=ospatsF_julia]{ospatsF_julia()}} \code{\link[=ospatsF_ref]{ospatsF_ref()}}
}
