% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospatsF-refactored.R
\name{ospatsF_ref}
\alias{ospatsF_ref}
\title{Legacy Ospats Algorithm}
\usage{
ospatsF_ref(
  x,
  covmodel_range,
  nstrata,
  niter = 100,
  niter_outer = 100,
  verbose = FALSE,
  temperature = 1,
  coolingrate = 0.95,
  rsquared = 1
)
}
\arguments{
\item{x}{data, with named columns (x, y, pred, var)}

\item{covmodel_range}{Exponential covariance model range parameter}

\item{nstrata}{number of strata}

\item{niter}{number of iterations of the allocation shuffle}

\item{niter_outer}{number of reruns of the allocation shuffle}

\item{rsquared}{adjustment R in the paper}
}
\description{
Refactored Ospats::ospatsF
}
