% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospatsMC.R
\name{ospatsMC}
\alias{ospatsMC}
\title{Ospats Algorithm Using Monte Carlo}
\usage{
ospatsMC(
  Z,
  nstrata,
  niter = 100,
  niter_outer = 100,
  verbose = 0,
  temperature = 1,
  coolingrate = 0.95,
  rsquared = 1,
  v = 1
)
}
\arguments{
\item{Z}{A matrix of n x M simulated values of the response}

\item{nstrata}{number of starta to consider}

\item{niter}{Number of iterations per one run of the optimisation algorithm}

\item{niter_outer}{Number of independent runs of the optimisation algorithm}

\item{verbose}{Print runtime diagnostics? (0,1,2)}

\item{temperature}{Annealing factor, will accept slightly bad moves with prob ~exp(-abs(delta)/temperature)}

\item{coolingrate}{Change temperature each interation by this factor. Should be at most 1.}

\item{rsquared}{The R^2 in the paper (default: 1)}
}
\description{
Ospats-algorithm with Monte Carlo samples.
}
\details{
Estimate the mean of the target using Monte Carlo sample of size M.

The initial stratification is based on cum-root-f, not k-means which is unreliable.
}
