% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospatsF-julia.R
\name{ospatsF_julia}
\alias{ospatsF_julia}
\title{Ospats Julia Version in R}
\usage{
ospatsF_julia(
  data,
  dRange,
  dRSquare,
  dMaxrun,
  nCycles,
  dStrata,
  interv,
  nmax,
  RSEmax = 1.5
)
}
\arguments{
\item{dat}{data, [x y pred var]}
}
\description{
Translate the Julia version of Ospats to R
}
\details{
The original Julia code:

https://github.com/jjdegruijter/ospats/ , file "ospatsmr"

The original code uses global variables; This version mimics the input naming of the original ospatsF.
}
